Adding 2 node sprite sheets
The default node is bunnfloat (when it's floating); I intend on adding
more sprites for when it hits the walls, but this is lower priority and
depends on my seriously deprived art skills.

The ghiblibunn is obviously based off of the ones in Ghibli films and
may be used for our bombs or some other node type.